Computes Lax pairs of systems of partial difference equations defined on quadrilaterals and consistent around a cube
A Mathematica package for the symbolic computation of Lax pairs of polynomials systems of nonlinear partial difference equations defined in quadrilaterals with the so-called `consistency around the cube' property. These are scalar PDeltaEs and systems of PDeltaEs studied by Adler, Bobenko, and Suris.

Written by Terry Bridgman and Willy Hereman.